Update PuzzleCreator
PuzzleCreator now creates a solution and a puzzle (with multiple
solutions).
For the moment, the puzzle creator selects a set amount of squares to
delete (based on difficulty) from random grid coordinates on the board.
Finding an algorithm to "dig holes" and keep the solution unique is
going to take a little longer than I had hoped. As far as functionality
is concerned however, there should be no problems with interfacing this
with the game class now