Create Game Feature
can now create a game of any difficulty by selection. The new game
button is just there as a tester and needs to be placed in its own
buttonpanel with the rest of the buttons at some point, but i didnt do
it myself as not to override any of Dave's changes. Just a small commit
so i might do more later tonight