Hint system
Added in the hint system:
it pretty much works of what you might call a point of interest system,
the player is altered to a cell that has a simple unique solution but
are left to find out what value should go in there. There are times in
expert matches where this is not the case for any cell, so ill need to
work something out there.